Node.js script for new United.com website and search
The new United.com uses XHR to fetch flight search results.
We can just used a standard node application to fetch these results
rather than the overhead of emulating the browser.  The XHR results
are parsed and displayed in a similar fashion.

The usage is: `node united.js SFO TPE 5/18/2016 5/20/2016`

Be sure to start with `npm install` to get all the packages

The benefit of this new approach is that all the requests can be
dispatched in parallel, saving significant time.
