+++ .gitignore
@@ -52,3 +52,6 @@ nosetests.xml
 .mr.developer.cfg
 .project
 .pydevproject
+
+# node
+node_modules/
