+++ united_request.json
@@ -1,0 +1,1 @@
+{"Revise":false,"UnaccompaniedMinorDisclamer":false,"searchTypeMain":"oneWay","Origin":"SFO","Destination":"TPE","DepartDate":"Jan 15, 2016","ReturnDate":"Jan 15, 2016","awardTravel":false,"MaxTrips":null,"numberOfTravelers":1,"numOfAdults":1,"numOfSeniors":0,"numOfChildren03":0,"numOfChildren02":0,"numOfChildren01":0,"numOfInfants":0,"numOfLapInfants":0,"travelerCount":1,"IsUnAccompaniedMinor":false,"MilitaryTravelType":null,"MilitaryOrGovernmentPersonnelStateCode":null,"tripLength":0,"flexMonth":null,"flexMonth2":null,"SortType":"bestmatches","cboMiles":"-1","cboMiles2":"-1","Trips":[{"DestinationAll":false,"returnARC":null,"connections":null,"nonStopOnly":false,"nonStop":true,"oneStop":true,"twoPlusStop":true,"DepartDate":"Jan 15, 2016","ReturnDate":null,"PetIsTraveling":false,"PreferredTime":"","PreferredTimeReturn":null,"Destination":"TPE","Index":1,"Origin":"SFO","Selected":false,"FormatedDepartDate":"Fri, Jan 15, 2016","OriginCorrection":null,"DestinationCorrection":null,"OriginAll":false}],"nonStopOnly":false,"CalendarOnly":false,"InitialShop":true,"IsSearchInjection":false,"CartId":"","CellIdSelected":null,"BBXSession":null,"SolutionSetId":null,"SimpleSearch":true,"RequeryForUpsell":false,"RequeryForPOSChange":false,"ShowClassOfServiceListPreference":false,"SelectableUpgradesOriginal":"MUA","RegionalPremierUpgradeBalance":0,"GlobalPremierUpgradeBalance":0,"RegionalPremierUpgrades":null,"GlobalPremierUpgrades":null,"FormattedAccountBalance":null,"GovType":null,"TripTypes":1,"flexible":false,"flexibleAward":false,"FlexibleDaysAfter":0,"FlexibleDaysBefore":0,"hiddenPreferredConn":null,"hiddenUnpreferredConn":null,"carrierPref":0,"chkFltOpt":0,"portOx":0,"travelwPet":0,"NumberOfPets":0,"cabinType":0,"cabinSelection":"ECONOMY","awardCabinType":0,"FareTypes":0,"FareWheelOnly":false,"buyUpgrade":0,"offerCode":null,"TVAOfferCodeLastName":null,"ClassofService":null,"UpgradeType":"MUA","BillingAddressCountryCode":null,"BillingAddressCountryDescription":null,"IsPassPlusFlex":false,"IsPassPlusSecure":false,"IsOffer":false,"IsMeetingWorks":false,"IsValidPromotion":false,"CalendarDateChange":null,"CoolAwardSpecials":false,"LastResultId":null,"IncludeLmx":false}
